+++ frequenz/resampling/__init__.py
@@ -6,6 +6,7 @@
 from ._rust_backend import (  # noqa: F401, F403 # pylint: disable=E0401
     Resampler,
     ResamplingFunction,
+    resample,
 )
 
-__all__ = ["Resampler", "ResamplingFunction"]
+__all__ = ["Resampler", "ResamplingFunction", "resample"]